Oscillator now pauses when no one is listening
Extraacts Painter (Terminal & Cursor) in separate file
