+++ lib/frontend/oscillator.dart
@@ -11,6 +11,21 @@ class Oscillator with Observable {
 
   var _value = true;
   Timer? _timer;
+  var _shouldRun = false;
+
+  @override
+  void addListener(listener) {
+    super.addListener(listener);
+    resume();
+  }
+
+  @override
+  void removeListener(listener) {
+    super.removeListener(listener);
+    if (listeners.isEmpty) {
+      pause();
+    }
+  }
 
   void _onOscillation(_) {
     _value = !_value;
@@ -22,11 +37,35 @@ class Oscillator with Observable {
   }
 
   void start() {
+    _shouldRun = true;
+    // only start right away when anyone is listening.
+    // the moment a listener gets registered the Oscillator will start
+    if (listeners.isNotEmpty) {
+      _startInternal();
+    }
+  }
+
+  void _startInternal() {
     if (_timer != null) return;
     _timer = Timer.periodic(duration, _onOscillation);
   }
 
+  void pause() {
+    _stopInternal();
+  }
+
+  void resume() {
+    if (_shouldRun) {
+      _startInternal();
+    }
+  }
+
   void stop() {
+    _shouldRun = false;
+    _stopInternal();
+  }
+
+  void _stopInternal() {
     _timer?.cancel();
     _timer = null;
   }
